# go/model_booking.go
/*
 * Travel Management service
 *
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * API version: 1.0.0
 * Generated by: Swagger Codegen (https://github.com/swagger-api/swagger-codegen.git)
 */
package swagger

type Booking struct {

	Id string `json:"id,omitempty"`

	EmpName string `json:"empName,omitempty"`

	From string `json:"from,omitempty"`

	Destination string `json:"destination,omitempty"`
}