# openapi.yaml
openapi: 3.0.1
info:
  title: Travel Management service
  version: 1.0.0
servers:
  - url: https://apis.preview-dv.choreo.dev/travelmgt/v1
security:
  - default: []
paths:
  /booking/{bookingId}:
    get:
      summary: Get booking by Id
      operationId: getBookigById
      parameters:
        - name: bookingId
          required: true
          in: path
          schema:
            type: string
      responses:
        200:
          description: Ok
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Booking'
  /booking:
    post:
      summary: Create booking
      operationId: createBooking
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BookingRequest'
      responses:
        201:
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Booking'
  /bookings:
    get:
      summary: >-
        Retrive all bookings.
      operationId: getBookings
      responses:
        '200':
          description: Ok
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Booking'
        '400':
          description: BadRequest
        '403':
          description: Forbidden
        '500':
          description: InternalServerError
      security:
        - default: []
      x-throttling-tier: Unlimited
      x-throttling-limit:
        requestCount: -1
        unit: MINUTE
      x-auth-type: Application & Application User
components:
  schemas:
    Booking:
      type: object
      properties:
        id: 
          type: string
        empName:
          type: string
        from:
          type: string
        destination:
          type: string
    BookingRequest:
      type: object
      properties:
        empId: 
         type: integer
        from:
          type: string
        destination:
          type: string
  securitySchemes:
    default:
      type: oauth2
      flows:
        implicit:
          authorizationUrl: https:///authorize
          scopes: {}
x-wso2-disable-security: 'false'